backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
openpyxl==3.1.5
